// src/model/organizationAuthenticationServerInformation.ts
/**
 * Stalker API
 * API di Stalker di Imola Informatica sviluppato da qbteam
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: qbteamswe@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Information of a user of an organizations.
 */
export interface OrganizationAuthenticationServerInformation { 
    /**
     * User unique identifier from the authentication server of the organization.
     */
    orgAuthServerId: string;
    /**
     * Name of the user.
     */
    name: string;
    /**
     * Surname of the user.
     */
    surname: string;
}